refactor: split monolithic main.rs (3264 lines) into focused modules
Extract CLI definitions, output formatting, and all command handlers
into dedicated modules under src/commands/ and src/. main.rs is now
~100 lines containing only mod declarations, main(), and run() dispatch.

New structure:
- src/cli.rs: Cli, Command, HooksAction structs
- src/output.rs: print_summary, print_coverage_*, print_*_markdown
- src/commands/mod.rs: shared helpers (load_and_discover, run_validation,
  compute_exit_code, exit_with_status, build_schema_columns) + tests
- src/commands/{check,coverage,generate,init,score,hooks,compact,
  resolve,diff,view,merge,issues,wizard,import}.rs: individual commands

Pure refactor with no behavioral changes. All 363 unit tests and 80
integration tests pass.

Closes #133
